backend/requirements.txt
# Core Django
Django>=5.0,<6.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0

# Database
psycopg2-binary>=2.9.9

# WebSocket & Real-time
channels>=4.0.0
channels-redis>=4.2.0

# Task Queue
celery>=5.3.0
redis>=5.0.0

# CORS
django-cors-headers>=4.3.0

# Production Server
gunicorn>=21.2.0
uvicorn[standard]>=0.24.0
daphne>=4.0.0

# Environment Variables (optional but recommended)
python-decouple>=3.8

# Monitoring & Error Tracking
sentry-sdk>=1.38.0

# Testing
pytest>=7.4.0
pytest-django>=4.7.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
factory-boy>=3.3.0

# Code Quality
black>=23.11.0
flake8>=6.1.0
isort>=5.12.0
mypy>=1.7.0

# Security
django-ratelimit>=4.1.0

# Cache (optional)
django-redis>=5.4.0

